requirements.txt
streamlit==1.38.0
openai==1.51.0
google-generativeai==0.8.2
yfinance==0.2.32
pandas==2.1.4
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
streamlit-chat==0.1.1
fredapi==0.5.1
numpy==1.24.3
datetime